% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr2.R
\name{create_diff}
\alias{create_diff}
\title{create a diff as character string}
\usage{
create_diff(oldFile = NULL, newFile = NULL)
}
\arguments{
\item{oldFile}{old}

\item{newFile}{new}
}
\description{
creates a diff of two text files. Based on jsdiff (BSD)
javascript library by Kevin Decker. If character vectors are supplied, a
diff will be created from them as well.
}
\examples{
 create_diff("foo\nbar", "foo\nbaz")

}
