% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr2.R
\name{diffr2Output}
\alias{diffr2Output}
\alias{renderDiffr2}
\title{Wrapper functions for using \pkg{diffr2} in \pkg{shiny}}
\usage{
diffr2Output(outputId, width = "100\%", height = "400px")

renderDiffr2(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{the width and height of the diffr2 widget}

\item{expr}{An expression that generates an HTML widget (or a
\href{https://rstudio.github.io/promises/}{promise} of an HTML widget).}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use \code{diffr2Output()} to create a UI element, and \code{renderDiffr2()}
to render the diffr2 widget.
}
\details{
When calling \code{renderDiffr2()} make sure that \code{divname}
matches the selected shiny output name. The div name is used to attach the
diff2html output. Otherwise the canvas will be blank.
}
