% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr2.R
\name{diffr2}
\alias{diffr2}
\title{View text file differences}
\usage{
diffr2(
  oldFile,
  newFile,
  diff,
  width = NULL,
  height = NULL,
  synchronisedScroll = TRUE,
  highlight = TRUE,
  fileListToggle = TRUE,
  fileListStartVisible = FALSE,
  fileContentToggle = TRUE,
  outputFormat = "line-by-line",
  drawFileList = TRUE,
  diffMaxChanges = NA,
  diffMaxLineLength = NA,
  matching = "none",
  matchWordsThreshold = 0.25,
  maxLineLengthHighlight = 10000,
  diffStyle = "word",
  renderNothingWhenEmpty = FALSE,
  matchingMaxComparisons = 2500,
  maxLineSizeInBlockForComparison = 200
)
}
\arguments{
\item{oldFile}{Your reference file (the old file).}

\item{newFile}{Your comparison file (the new file).}

\item{diff}{Alternatively you can provide your diff file. If provided, it overrides oldFile & newFile diff.}

\item{width}{for \code{\link{createWidget}}}

\item{height}{for \code{\link{createWidget}}}

\item{synchronisedScroll}{scroll both panes in side-by-side mode: TRUE or , default is TRUE}

\item{highlight}{syntax highlight the code on the diff: TRUE or FALSE, default is TRUE}

\item{fileListToggle}{allow the file summary list to be toggled: TRUE or FALSE, default is TRUE}

\item{fileListStartVisible}{choose if the file summary list starts visible: TRUE or FALSE, default is FALSE}

\item{fileContentToggle}{allow each file contents to be toggled: TRUE or FALSE, default is TRUE}

\item{outputFormat}{the format of the output data: 'line-by-line' or 'side-by-side', default is 'line-by-line'}

\item{drawFileList}{show a file list before the diff: TRUE or FALSE, default is TRUE}

\item{diffMaxChanges}{number of changed lines after which a file diff is deemed as too big and not displayed, default is undefined}

\item{diffMaxLineLength}{number of characters in a diff line after which a file diff is deemed as too big and not displayed, default is undefined}

\item{matching}{matching level: 'lines' for matching lines, 'words' for matching lines and words or 'none', default is none}

\item{matchWordsThreshold}{similarity threshold for word matching, default is 0.25}

\item{maxLineLengthHighlight}{only perform diff changes highlight if lines are smaller than this, default is 10000}

\item{diffStyle}{show differences level in each line: 'word' or 'char', default is 'word'}

\item{renderNothingWhenEmpty}{render nothing if the diff shows no change in its comparison: TRUE or FALSE, default is FALSE}

\item{matchingMaxComparisons}{perform at most this much comparisons for line matching a block of changes, default is 2500}

\item{maxLineSizeInBlockForComparison}{maximum number os characters of the bigger line in a block to apply comparison, default is 200}
}
\description{
Shows diff based on the diff2html (MIT) js library by Rodrigo Fernandes
}
\examples{
library(diffr2)
file1 = tempfile()
writeLines("hello, world!\n", con = file1)
file2 = tempfile()
writeLines(paste0(
"hello world?\nI don't get it\n",
paste0(sample(letters, 65, replace = TRUE), collapse = "")), con = file2)
diffr2(file1, file2)
}
