% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffr2.R
\name{create_diff}
\alias{create_diff}
\title{create the diff as character string}
\usage{
create_diff(oldFile, newFile)
}
\arguments{
\item{oldFile}{old}

\item{newFile}{new}
}
\description{
shows the diff of two text files. Based on diffjs (BSD) js library by Kevin Decker
}
